+++ tests/HighTests/test/clients/prestashop-client.js
@@ -1,0 +1,38 @@
+const {getClient} = require('../common.webdriverio.js');
+const {selector} = require('../globals.webdriverio.js');
+
+class PrestashopClient {
+  constructor() {
+    this.client = getClient();
+  }
+
+  signinBO() {
+    return this.client.signinBO();
+  }
+
+  signoutBO() {
+    return this.client.signoutBO();
+  }
+
+  signinFO() {
+    return this.client.signinFO();
+  }
+
+  signoutFO() {
+    return this.client.signoutFO();
+  }
+
+  takeScreenshot() {
+    return this.client.saveScreenshot(`screenshots/${this.client.desiredCapabilities.browserName}_exception_${global.date_time}.png`);
+  }
+
+  open() {
+    return this.client.init().windowHandleSize({width: 1280, height: 1024});
+  }
+
+  close() {
+    return this.client.end();
+  }
+}
+
+module.exports = PrestashopClient;
