+++ tests/HighTests/test/mocha.opts
@@ -1,0 +1,4 @@
+--require ./test/helpers.js
+--timeout 200000
+--slow 45000
+--reporter mochawesome
